Enum sizes (closes #156) (#169)
* Parse enum sizes

* More checking, start C# and Rust codegen

* Fix BigInteger parsing and C# tests

* Rust test + fixes

* C++ test + fix

* TypeScript 64-bit enum special-case + test

* Proper enum size read/write in TypeScript + test

* Add codegen step to build-runtime-cs.yml

* Update build-runtime-cs.yml

* Generate TypeScript enums both ways

* Better error messages and function names and docs!

+++ Runtime/Rust/src/serialization/error.rs
@@ -10,7 +10,7 @@ pub enum DeserializeError {
     CorruptFrame,
     /// There was an issue with a string encoding
     Utf8EncodingError(std::str::Utf8Error),
-    InvalidEnumDiscriminator(u32),
+    InvalidEnumDiscriminator(i128),
     /// A message type had multiple definitions for the same field
     DuplicateMessageField,
 }
